[PT Run] Moved BrowserInfo.cs to Wox.Plugin/SharedCommands and removed Plugins/Common
