+++ src/modules/launcher/Plugins/Community.PowerToys.Run.Plugin.WebSearch/Community.PowerToys.Run.Plugin.WebSearch.csproj
@@ -102,11 +102,5 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-  
-  <ItemGroup>
-    <Compile Include="..\Common\BrowserInfo.cs">
-      <Link>BrowserInfo.cs</Link>
-    </Compile>
-  </ItemGroup>
 
 </Project>
